+++ CHANGES.rst
@@ -6,6 +6,7 @@ Skeem changelog
 in progress
 ===========
 - Added support for Python 3.12
+- Adjusted SQL DDL for sqlalchemy-cratedb 0.40.0
 
 
 2023-03-09 0.1.0