+++ CHANGES.rst
@@ -7,6 +7,7 @@ in progress
 ===========
 - Added support for Python 3.12
 - Adjusted SQL DDL for sqlalchemy-cratedb 0.40.0
+- Adjusted ddlgenerator wrapper for pandas 2
 
 
 2023-03-09 0.1.0