+++ CHANGES.rst
@@ -5,7 +5,7 @@ Skeem changelog
 
 in progress
 ===========
-- Added support for Python 3.12
+- Added support for Python 3.12 and 3.13
 - Adjusted SQL DDL for sqlalchemy-cratedb 0.40.0
 - Adjusted ddlgenerator wrapper for pandas 2