+++ CHANGES.rst
@@ -5,14 +5,16 @@ Skeem changelog
 
 in progress
 ===========
+
+2024-10-22 v0.1.1
+=================
 - Added support for Python 3.12 and 3.13
 - Adjusted SQL DDL for sqlalchemy-cratedb 0.40.0
 - Adjusted ddlgenerator wrapper for pandas 2
 - Updated to frictionless >=5.16
 
-
-2023-03-09 0.1.0
-================
+2023-03-09 v0.1.0
+=================
 - Initial working version, supporting JSON
 - Add CLI subsystem
 - Add logging